add another uploader for profile photo that has public read
