+++ app/uploaders/profile_photo_attachment_uploader.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class ProfilePhotoAttachmentUploader < CarrierWave::Uploader::Base
-  PROCESSING_CLASS = PhotoProcessingUploader
+  PROCESSING_CLASS = VIC::PhotoProcessingUploader
   include ValidateFileSize
   include SetAwsConfig
   include AsyncProcessing
