+++ app/uploaders/vic/photo_processing_uploader.rb
@@ -1,0 +1,11 @@
+module VIC
+  class PhotoProcessingUploader < ProcessingUploader
+    def initialize(*args)
+      super
+
+      if Rails.env.production?
+        self.aws_acl = 'public-read'
+      end
+    end
+  end
+end
